Updating from JCR API to Sling API
I want to use the AssetManager and Asset interfaces, so I need to use
resourceResolver. This is a rewrite to make that happen.

The initialize is finished, now I just need to update the sync methods.

I also updated the Service UserMapperImpl so that it actually used my
evernote user